Moved interface to interface directory, removed commented out code, moved function to helper utils.
