+++ services/common/src/interfaces/verifiableCredentials/index.ts
@@ -1,2 +1,3 @@
 export * from "./verifiableCredentialInvitation.interface";
 export * from "./credentialExchangeDetail.interface";
+export * from "./DigitalCredentialsPermit.interface";
